Create WDI correlations matrix.py
The WDI CSV bundle is a 42.5-Mbyte zipped archive. After downloading and unzipping it, you’ll see that the main file is called WDI_Data.csv. A good way to get an overview of the file contents is to examine it interactively. Because we’ll be using Python, the best way to interact with the tools that we’ll use is by launching a session of iPython, and then loading the data:
In [1]: import pandas as pd
In [2]: data = pd.read_csv(“WDI_Data.csv”)
The result, in data, is a DataFrame containing the data. Think of a DataFrame as a two-dimensional array with some extra features that allow for easy manipulation. In a DataFrame, data is organized in columns and an index (corresponding to the rows). If we enter
In [3]: data.columns
we’ll get an output that shows the names of the columns: the country name, the code for that country, an indicator name, and an indicator code. These are followed by columns for each year from 1960 to 2012. Similarly, if we enter
In [4]: data.index
we’ll see that the data contains 352409 rows. Each row corresponds to the values of one particular indicator for one country for the years 1960 to 2014; years without values in a row indicate no measurement in that year for that indicator in that country. Let’s see first, how many indicators there are
In [5]: len(data[‘Indicator Name’].unique())
Out[5]: 1289
and second, how many countries there are
In [6]: len(data[‘Country Name’].unique())
Out[6]: 246
Now we have a problem to solve: Are the indicators independent among themselves, or are some of them related to others?
Because we measure indicators by year and by country, we must more precisely define the problem by de-ciding which parameters to retain as constant. In general, we get the best statistical results as our samples increase. It makes sense then to rephrase the problem: For the year in which we have most measurements, are the most measured indicators independent among themselves, or as some of them related to others? By “most measured indicators,” we mean those that have been measured in more countries. It turns out that we can find the answer to the question in about 50 LOC. Figure 1 contains the full program.
Lines 1–6 are imports of the libraries that we’ll be using. Line 7 reads the data. In line 8, we give the number of most measured indicators that we would like to examine. In line 10, we declared array from 1960 to 2014 becuase the data in file has only filled data with upto 2014 and in line9 we find the zero-based position of the first column with yearly measurements. After that, we’re able in line 11 to find the column with the most measurements (the year 2010). We then remove all data for which measurements aren’t available. In lines 13-20, we get the most measured indicators.
The actual statistical calculations start from line 21, where we prepare a table of ones to hold the result of the correlation values between each pair of indicators. In the loop that follows, we calculate each pairwise correlation and store it in the table we prepared. Finally, in lines 32-43, we display the results on screen and save them to a PDF file.We take care to reverse the vertical order of the correlation matrix so that the most important indicator comes on the top of the matrix (lines 33 and 42).
The diagonal has perfect correlation—as it should, because we’re examining the same indicators. In addition to that, we do see that there are indicators that correlate with each other—some positively, even strongly so, and some negatively or very negatively.